+++ fuel-asm/src/lib.rs
@@ -183,12 +183,14 @@ impl_instructions! {
     0x3E ECK1 eck1 [RegId RegId RegId]
     "The 64-byte Secp256r1 public key (x, y) recovered from 64-byte signature on 32-byte message."
     0x3F ECR1 ecr1 [RegId RegId RegId]
+    "Verify ED25519 public key and signature match a 32-byte message."
+    0x40 ED19 ed19 [RegId RegId RegId]
     "The keccak-256 hash of a slice."
-    0x40 K256 k256 [RegId RegId RegId]
+    0x41 K256 k256 [RegId RegId RegId]
     "The SHA-2-256 hash of a slice."
-    0x41 S256 s256 [RegId RegId RegId]
+    0x42 S256 s256 [RegId RegId RegId]
     "Get timestamp of block at given height."
-    0x42 TIME time [RegId RegId]
+    0x43 TIME time [RegId RegId]
 
     "Performs no operation."
     0x47 NOOP noop []
@@ -480,9 +482,9 @@ impl Opcode {
         match self {
             ADD | AND | DIV | EQ | EXP | GT | LT | MLOG | MROO | MOD | MOVE | MUL | NOT | OR | SLL | SRL | SUB
             | XOR | WDCM | WQCM | WDOP | WQOP | WDML | WQML | WDDV | WQDV | WDMD | WQMD | WDAM | WQAM | WDMM | WQMM
-            | RET | ALOC | MCL | MCP | MEQ | ECK1 | ECR1 | K256 | S256 | NOOP | FLAG | ADDI | ANDI | DIVI | EXPI
-            | MODI | MULI | MLDV | ORI | SLLI | SRLI | SUBI | XORI | JNEI | LB | LW | SB | SW | MCPI | MCLI | GM
-            | MOVI | JNZI | JI | JMP | JNE | JMPF | JMPB | JNZF | JNZB | JNEF | JNEB | CFEI | CFSI | CFE | CFS
+            | RET | ALOC | MCL | MCP | MEQ | ECK1 | ECR1 | ED19 | K256 | S256 | NOOP | FLAG | ADDI | ANDI | DIVI
+            | EXPI | MODI | MULI | MLDV | ORI | SLLI | SRLI | SUBI | XORI | JNEI | LB | LW | SB | SW | MCPI | MCLI
+            | GM | MOVI | JNZI | JI | JMP | JNE | JMPF | JMPB | JNZF | JNZB | JNEF | JNEB | CFEI | CFSI | CFE | CFS
             | GTF => true,
             _ => false,
         }