+++ fuel-vm/src/interpreter/crypto.rs
@@ -1,14 +1,12 @@
 use super::internal::{clear_err, inc_pc, set_err};
-use super::memory::{try_mem_write, try_zeroize, OwnershipRegisters};
+use super::memory::{read_bytes, try_mem_write, try_zeroize, OwnershipRegisters};
 use super::{ExecutableTransaction, Interpreter};
 use crate::constraints::reg_key::*;
 use crate::consts::{MEM_MAX_ACCESS_SIZE, MEM_SIZE, MIN_VM_MAX_RAM_USIZE_MAX, VM_MAX_RAM};
 use crate::error::RuntimeError;
 
 use crate::arith::{checked_add_word, checked_sub_word};
-use crate::prelude::MemoryRange;
 use fuel_asm::PanicReason;
-use fuel_crypto::coins_bip32::prelude::RecoveryId;
 use fuel_crypto::{Hasher, Message, PublicKey, Signature};
 use fuel_types::{Bytes32, Bytes64, Word};
 
@@ -29,6 +27,11 @@ where
         let owner = self.ownership_registers();
         let (SystemRegisters { err, pc, .. }, _) = split_registers(&mut self.registers);
         secp256r1_recover(&mut self.memory, owner, err, pc, a, b, c)
+    }
+
+    pub(crate) fn ed25519_verify(&mut self, a: Word, b: Word, c: Word) -> Result<(), RuntimeError> {
+        let (SystemRegisters { err, pc, .. }, _) = split_registers(&mut self.registers);
+        ed25519_verify(&mut self.memory, err, pc, a, b, c)
     }
 
     pub(crate) fn keccak256(&mut self, a: Word, b: Word, c: Word) -> Result<(), RuntimeError> {
@@ -92,31 +95,41 @@ pub(crate) fn secp256r1_recover(
     b: Word,
     c: Word,
 ) -> Result<(), RuntimeError> {
-    use p256::ecdsa::{Signature, VerifyingKey};
+    let sig = Bytes64::from(read_bytes(memory, b)?);
+    let msg = Bytes32::from(read_bytes(memory, c)?);
 
-    let sig = MemoryRange::new(b, Bytes64::LEN)?;
-    let msg = MemoryRange::new(c, Bytes32::LEN)?;
-
-    let Ok(signature) = Signature::from_slice(&memory[sig.usizes()]) else {
-        // Invalid signature, so there is no valid key to recover
-        try_zeroize(a as usize, PublicKey::LEN, owner, memory)?;
-        set_err(err);
-        return inc_pc(pc);
-    };
-
-    // Attempt the four possible recovery ids
-    for recid in 0..RecoveryId::MAX {
-        let recid = RecoveryId::from_byte(recid).unwrap();
-        if let Ok(pub_key) = VerifyingKey::recover_from_msg(&memory[msg.usizes()], &signature, recid) {
-            try_mem_write(a, &pub_key.to_sec1_bytes(), owner, memory)?;
+    match fuel_crypto::secp256r1::recover(&sig, &msg) {
+        Ok(pub_key) => {
+            try_mem_write(a, &pub_key, owner, memory)?;
             clear_err(err);
-            return inc_pc(pc);
+        }
+        Err(_) => {
+            try_zeroize(a as usize, PublicKey::LEN, owner, memory)?;
+            set_err(err);
         }
     }
 
-    // None of the keys matched
-    try_zeroize(a as usize, PublicKey::LEN, owner, memory)?;
-    set_err(err);
+    inc_pc(pc)
+}
+
+pub(crate) fn ed25519_verify(
+    memory: &mut [u8; MEM_SIZE],
+    err: RegMut<ERR>,
+    pc: RegMut<PC>,
+    a: Word,
+    b: Word,
+    c: Word,
+) -> Result<(), RuntimeError> {
+    let pub_key = Bytes32::from(read_bytes(memory, a)?);
+    let sig = Bytes64::from(read_bytes(memory, b)?);
+    let msg = Bytes32::from(read_bytes(memory, c)?);
+
+    if fuel_crypto::ed25519::verify(&pub_key, &sig, &msg).is_ok() {
+        clear_err(err);
+    } else {
+        set_err(err);
+    }
+
     inc_pc(pc)
 }