+++ fuel-vm/src/tests/crypto.rs
@@ -1,6 +1,6 @@
 use fuel_asm::PanicReason::{ArithmeticOverflow, ErrorFlag, MemoryOverflow};
 use fuel_asm::{op, GTFArgs, RegId};
-use fuel_crypto::Hasher;
+use fuel_crypto::{Hasher, PublicKey, SecretKey, Signature};
 use fuel_tx::TransactionBuilder;
 use rand::rngs::StdRng;
 use rand::SeedableRng;